Created ArgumentParser, as well as ArgumentParseException
